Add pip install support
