+++ .github/workflows/build-macast.yaml
@@ -83,10 +83,10 @@ jobs:
             'pip install -r requirements/common.txt && \
             pyinstaller --noconfirm -F -w \
               --additional-hooks-dir=. \
-              --add-data=".version:." \
+              --add-data="macast/.version:macast/." \
               --add-data="macast/xml/*:macast/xml" \
               --add-data="i18n/zh_CN/LC_MESSAGES/*.mo:i18n/zh_CN/LC_MESSAGES" \
-              --add-data="assets/*.png:assets" \
+              --add-data="macast/assets/*.png:macast/assets" \
               --exclude-module=tkinter \
               --distpath="app" \
             Macast.py'
@@ -151,7 +151,7 @@ jobs:
         echo "VERSION=v${VERSION}" | Out-File -FilePath $env:GITHUB_ENV -Encoding utf8 -Append
         pyinstaller --noconfirm -F -w --additional-hooks-dir=. --add-data=".version;." --add-data="macast/xml/*;macast/xml"  --add-data="i18n/zh_CN/LC_MESSAGES/*.mo;i18n/zh_CN/LC_MESSAGES" --add-data="assets/*;assets" --add-binary="bin/mpv.exe;bin" --icon=assets/icon.ico Macast.py
         cp dist/Macast.exe Macast-Windows-v${VERSION}.exe
-        pyinstaller --noconfirm -F --additional-hooks-dir=. --add-data=".version;." --add-data="macast/xml/*;macast/xml"  --add-data="i18n/zh_CN/LC_MESSAGES/*.mo;i18n/zh_CN/LC_MESSAGES" --add-data="assets/*;assets" --add-binary="bin/mpv.exe;bin" --icon=assets/icon.ico Macast.py
+        pyinstaller --noconfirm -F --additional-hooks-dir=. --add-data="macast/.version;macast/." --add-data="macast/xml/*;macast/xml"  --add-data="i18n/zh_CN/LC_MESSAGES/*.mo;i18n/zh_CN/LC_MESSAGES" --add-data="macast/assets/*;macast/assets" --add-binary="bin/mpv.exe;bin" --icon=macast/assets/icon.ico Macast.py
         cp dist/Macast.exe Macast-Windows-v${VERSION}-debug.exe
         echo "::set-output name=status::success"
 
@@ -186,7 +186,7 @@ jobs:
         wget https://laboratory.stolendata.net/~djinn/mpv_osx/mpv-latest.tar.gz
         mkdir -p bin && tar --strip-components 2 -C bin -xzvf mpv-latest.tar.gz mpv.app/Contents/MacOS
         msgfmt -o i18n/zh_CN/LC_MESSAGES/macast.mo i18n/zh_CN/LC_MESSAGES/macast.po
-        python3 setup.py py2app
+        python3 setup_py2app.py py2app
         cp -R bin dist/Macast.app/Contents/Resources/
         export VERSION=`cat .version`
         echo "DIST_DMG=Macast-MacOS-v${VERSION}.dmg" >> $GITHUB_ENV
