+++ macast/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2021 by xfangfang. All Rights Reserved.
 
 
-from .macast import Service
+from .macast import Service, Macast, gui, cli
 from .utils import SETTING_DIR, Setting, SettingProperty
 from .renderer import Renderer, RendererSetting
 from .plugin import RendererPlugin
