+++ setup.py
@@ -6,31 +6,75 @@ Usage:
 """
 
 import os
-from setuptools import setup
+import sys
+from setuptools import setup, find_packages
 
-APP = ['Macast.py']
-DATA_FILES = ['.version', 'i18n', 'assets']
 VERSION = "0.0.0"
-with open('.version', 'r') as f: VERSION = f.read().strip()
-OPTIONS = {
-    'argv_emulation': True,
-    'plist': {
-        'LSUIElement': True,
-        'NSHighResolutionCapable': True,
-        'LSMinimumSystemVersion': '10.15.0',
-        'CFBundleIdentifier': 'cn.xfangfang.Macast',
-        'NSHumanReadableCopyright': 'Copyright 2021 xfangfang and the Macast contributors.',
-        'CFBundleShortVersionString': str(VERSION),
-        'CFBundleVersion': str(VERSION),
-    },
-    'packages': ['rumps', 'macast'],
-    'iconfile': os.path.abspath('assets/icon.icns')
-}
+with open('macast/.version', 'r') as f:
+    VERSION = f.read().strip()
+LONG_DESCRIPTION = ""
+with open('README.md', 'r') as f:
+    LONG_DESCRIPTION = f.read()
+OPTIONS = {}
+INSTALL = ["requests", "appdirs", "cherrypy", "lxml"]
+PACKAGES = find_packages()
+DATA_FILES = [('xml',
+               ['macast/xml/SinkProtocolInfo.csv',
+                'macast/xml/AVTransport.xml',
+                'macast/xml/ConnectionManager.xml',
+                'macast/xml/Description.xml',
+                'macast/xml/RenderingControl.xml'
+                ]),
+              ('assets',
+               ["macast/assets/icon.icns",
+                "macast/assets/icon.ico",
+                "macast/assets/icon.png",
+                "macast/assets/menu_dark_large.png",
+                "macast/assets/menu_dark.png",
+                "macast/assets/menu_light_large.png",
+                "macast/assets/menu_light.png"
+                ]),
+              ('',
+               ['macast/.version',
+                ])
+              ]
+
+if sys.platform == 'darwin':
+    INSTALL.append("rumps")
+else:
+    INSTALL.append("pillow",
+                   "git+https://github.com/xfangfang/pystray.git",
+                   "git+https://github.com/xfangfang/pyperclip.git")
 
 setup(
-    app=APP,
+    name="macast",
+    version=VERSION,
+    author="xfangfang",
+    author_email="xfangfang@126.com",
+    description="a DLNA Media Renderer",
+    license="GPL3",
+    url="https://github.com/xfangfang/Macast",
+    long_description=LONG_DESCRIPTION,
+    long_description_content_type="text/markdown",
+    classifiers=["Topic :: Multimedia :: Sound/Audio",
+                 "Topic :: Multimedia :: Video",
+                 'Programming Language :: Python :: 3',
+                 'Programming Language :: Python :: 3.6',
+                 'Programming Language :: Python :: 3.7',
+                 'Programming Language :: Python :: 3.8',
+                 'Programming Language :: Python :: 3.9',
+                 ],
+    keywords=["mpv", "dlna", "renderer"],
+    options=OPTIONS,
+    install_requires=INSTALL,
+    packages=PACKAGES,
     data_files=DATA_FILES,
-    options={'py2app': OPTIONS},
-    setup_requires=['py2app'],
-    py_modules=[]
+    include_package_data=True,
+    entry_points={
+        'console_scripts': [
+            'macast-cli = macast.macast:cli',
+            'macast-gui = macast.macast:gui'
+        ]
+    },
+    python_requires=">=3.6",
 )
