+++ setup_py2app.py
@@ -1,0 +1,37 @@
+"""
+This is a setup.py script generated by py2applet
+
+Usage:
+    python setup.py py2app
+"""
+
+import os
+from setuptools import setup
+
+APP = ['Macast.py']
+DATA_FILES = ['macast/.version', 'i18n', 'macast/assets']
+VERSION = "0.0.0"
+with open('macast/.version', 'r') as f:
+    VERSION = f.read().strip()
+OPTIONS = {
+    'argv_emulation': True,
+    'plist': {
+        'LSUIElement': True,
+        'NSHighResolutionCapable': True,
+        'LSMinimumSystemVersion': '10.15.0',
+        'CFBundleIdentifier': 'cn.xfangfang.Macast',
+        'NSHumanReadableCopyright': 'Copyright 2021 xfangfang and the Macast contributors.',
+        'CFBundleShortVersionString': str(VERSION),
+        'CFBundleVersion': str(VERSION),
+    },
+    'packages': ['rumps', 'macast'],
+    'iconfile': os.path.abspath('macast/assets/icon.icns')
+}
+
+setup(
+    app=APP,
+    data_files=DATA_FILES,
+    options={'py2app': OPTIONS},
+    setup_requires=['py2app'],
+    py_modules=[]
+)
